New CircleCI job e2e "need fixing"
Split often failing end to end tests to separate "need fixing" category. We can
set GitHub to not require those tests to pass for merging. Of course we should
dedicate some effort to actually fix these tests, but this avoids having to use
admin privileges to merge the PR.
